# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fitness Tracking Platform API
  version: 1.0.0
  description: Authoritative REST API for multi-user fitness management, daily tracking, workout/diet plan builders, progress analysis, reminders, and administration.
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
paths:
...
  /me/meals/today:
    get:
      summary: Get today diet meals with configured food options
      tags: [Meals]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Today meals
  /me/meals/history:
    get:
      summary: Get historical meal logs
      tags: [Meals]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          required: false
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Meal history
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { type: object }
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /me/meals/{mealId}/log:
    put:
      summary: Log/update meal completion with food option selections
      tags: [Meals]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: mealId
          in: path
          required: true
          schema: { type: integer }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                logDate: { type: string, format: date }
                status: { type: string, enum: [completed, skipped, partial] }
                notes: { type: string }
                selections:
                  type: array
                  items:
                    type: object
                    required: [optionGroupId, optionId]
                    properties:
                      optionGroupId: { type: integer }
                      optionId: { type: integer }
                      foodId: { type: integer }
                      quantity: { type: number, exclusiveMinimum: 0 }
                clientOperationId: { type: string }
      responses:
        '200':
          description: Meal logged
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Pagination:
      type: object
      required: [page, limit, total, totalPages]
      properties:
        page: { type: integer, minimum: 1 }
        limit: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...